requirements.txt
--upgrade pip
flask==2.0.1
python-dotenv==0.19.0
autogen==1.0.0
gunicorn==20.1.0